+++ js/util.js
@@ -1,0 +1,16 @@
+const getRandomNumber = function (min, max) {
+  if (min < max) {
+    return Math.floor(Math.random() * (max + 1 - min) + min);
+  }
+  alert('Минимальное значение не может быть больше максимального!');
+}
+
+const checkStringLength = function (string, maxLength) {
+  return string.length <= maxLength;
+}
+
+function getRandomArrayElement(array) {
+  return array[getRandomNumber(0, array.length - 1)];
+}
+
+export {getRandomNumber, checkStringLength, getRandomArrayElement};
